Reduce precision sent to roshi to match db.
This ensures that roshi only cares about microseconds and nanoseconds
are discarded. This is important because postgres only stores
microsecond precision. In cases where we write to the stream from an
object recently created in ruby memory we have nanosecond precision.

Due to the nature of roshi adding/removal of items from streams has a
"highest score wins" approach. In our case ts is our score.

This fix ensures that all writes to streams can be removed, as both
addition and removal will be the same timestamp.